refactoring The code
