+++ src/index.html
@@ -14,24 +14,19 @@
     <main>
       <header class="header">
         <a href="#"
-          ><img src="./assets/komikult.png" alt="komikult" class="logo"
+          ><img src="../src/assets/komikult.png" alt="komikult" class="logo"
         /></a>
         <nav class="nav-bar">
           <ul class="nav-items">
             <li class="nav-item"><a href="">Home</a></li>
-            <li class="nav-item"><a href="">Heroes</a></li>
+            <li class="nav-item"><a href="">Heroes</a> (<span class="items-number"></span>)</li>
             <li class="nav-item"><a href="">About</a></li>
-            <li class="nav-item items-number"></li>
           </ul>
         </nav>
       </header>
       <section class="charct" id="charct">
         <div class="container">
           <ul class="charc-list" id="charc-list"></ul>
-
-          <div class="modal" id="modal"></div>
-        </div>
-        <div id="overlay"></div>
         </div>
       </section>
     </main>
