--- src/js/dom.js
@@ -1,41 +1,0 @@
-// import { Api } from "./api";
-// import {getLikesNumber } from "./likesApi";
-
-
-// const charactersList = document.getElementById('charc-list');
-
-// class DomElement {
-//   static addCharacterList = async () => {
-    // let names, images, description,id,likeNum;
-    // const marvelData = await Api.getCharacters()
-  
-    // for (let i = 0; i < 20 ; i += 1) {
-      // targeting the name and img url
-      // names =  marvelData.data.results[i].name;
-      // id =  marvelData.data.results[i].id;
-      // images =  marvelData.data.results[i].thumbnail;
-      // description =  marvelData.data.results[i].description;
-      // creating the img url
-      // let imageUrl = `${images.path}/portrait_incredible.${images.extension}`;
-      // creating list, header and img elements
-      
-      
-  //     charactersList.innerHTML += `
-  //     <li class="char-item" id="${id}">
-  //       <img class="char-img" src="${imageUrl}">
-  //       <h2 class="char-name">${names}</h2>
-  //       <button class="like">Like</button><span class="likesCounter">${likeNum}</span>
-  //       <button data-modal-target="#modal" class="comment">Comment</button>
-  //     </li>
-  //     `;
-  //   }
-  //  }
-
-
-
-//  export {DomElement}
-
-
-
-
-
